New consumer for pagure over dist-git.
Through conversations with @pypingou, we identified that this thingie
needed a new consumer to write out a gitolite header for pagure (one
that defines the provenpackager and other admin groups for git access).

This should do that.

The gitolite header gets written out to some location.  Pagure will have
to be responsible for pulling it in and combining it with the real
gitolite config.